add_setup_py: Support for creating setup.py
setuptools/setup.py is a useful, but peculiar, tool for building python
packages that is well integration with PyPI. It's a build system of its
own, and for some projects that essentially means double bookeeping of
build scripts.

CMake has a lot of information on what is to be built and how, so it's
fairly natural to want to use this build information to create a
setup.py that can build and upload this package as a step in an
automated build+deploy process.

add_setup_py does exactly this, assuming add_python_package is used and
the base cmakelists is reasonably well set up.