Filter and rename Windows paths/.defs
Colons are filtered out of path names, since it's illegal for them to
appear anywhere not immediately after disk labels. This is rarely a
problem on *nix, but shows up all the time on Windows.

Source files given to a target that end in .def are for exporting
symbols from libraries, but this isn't interesting when compiling all
sources into the same object. These files can also make some compilers
choke, so its better to filter them out.